+++ server/src/main/java/replicant/server/transport/ReplicantMessageBrokerImpl.java
@@ -1,12 +1,13 @@
 package replicant.server.transport;
 
 import java.util.Collection;
+import java.util.HashSet;
 import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.LinkedBlockingDeque;
-import java.util.concurrent.ScheduledFuture;
+import java.util.concurrent.LinkedBlockingQueue;
 import java.util.concurrent.TimeUnit;
-import java.util.concurrent.locks.ReentrantLock;
+import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.concurrent.atomic.AtomicInteger;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 import javax.annotation.Nonnull;
@@ -20,6 +21,7 @@ import javax.enterprise.inject.Typed;
 import javax.inject.Inject;
 import javax.json.JsonValue;
 import javax.transaction.Transactional;
+import javax.websocket.CloseReason;
 import org.jetbrains.annotations.VisibleForTesting;
 import replicant.server.ChangeSet;
 import replicant.server.EntityMessage;
@@ -32,37 +34,32 @@ public class ReplicantMessageBrokerImpl
 {
   @Nonnull
   private static final Logger LOG = Logger.getLogger( ReplicantMessageBrokerImpl.class.getName() );
-  private static final long QUEUE_TIMEOUT = 10L;
+  private static final long RETRY_DELAY = 20L;
   @Nonnull
-  private final BlockingQueue<ReplicantSession> _queue = new LinkedBlockingDeque<>();
-  /**
-   * In progress is used because a single session with a long running load can queue up other requests
-   * and will eventually consume up all the active processors. If we guarantee that there is at most
-   * one thread per session then we can allow sessions to keep making progress.
-   */
+  private final BlockingQueue<ReplicantSession> _queue = new LinkedBlockingQueue<>();
   @Nonnull
-  private final ConcurrentHashMap<String, ReplicantSession> _inProgress = new ConcurrentHashMap<>();
+  private final ConcurrentHashMap<String, WorkState> _workStates = new ConcurrentHashMap<>();
+  @Nonnull
+  private final AtomicInteger _activeDrainTasks = new AtomicInteger();
+  @Nonnull
+  private final AtomicBoolean _retryScheduled = new AtomicBoolean();
   @VisibleForTesting
   @Inject
   ReplicantSessionManager _sessionManager;
   @Resource( lookup = "java:comp/DefaultManagedScheduledExecutorService" )
   private ManagedScheduledExecutorService _executor;
-  private ScheduledFuture<?> _future;
-
-  @PostConstruct
-  void postConstruct()
-  {
-    _future = _executor.scheduleAtFixedRate( this::processPendingSessions, 3, 20, TimeUnit.MILLISECONDS );
-  }
+  @Resource( lookup = "replicant/broker/maxConcurrentDrainTasks" )
+  private Integer _maxConcurrentDrainTasks;
+  @Resource( lookup = "replicant/broker/maxPacketsPerRun" )
+  private Integer _maxPacketsPerRun;
+  @Resource( lookup = "replicant/broker/maxSessionsPerDrainTask" )
+  private Integer _maxSessionsPerDrainTask;
+  private volatile boolean _stopping;
 
   @PreDestroy
   void preDestroy()
   {
-    if ( null != _future )
-    {
-      _future.cancel( true );
-      _future = null;
-    }
+    _stopping = true;
   }
 
   @Nonnull
@@ -77,7 +74,8 @@ public class ReplicantMessageBrokerImpl
   {
     final var packet = new Packet( altersExplicitSubscriptions, requestId, response, etag, messages, changeSet );
     session.queuePacket( packet );
-    _queue.add( session );
+    enqueueSessionIfRequired( session );
+    scheduleDrainTasks();
     if ( LOG.isLoggable( Level.FINEST ) )
     {
       LOG.log( Level.FINEST, () -> "queueChangeMessage() queue.size=" + _queue.size() + " packet=" + packet );
@@ -85,59 +83,288 @@ public class ReplicantMessageBrokerImpl
     return packet;
   }
 
-  @Override
-  public void processPendingSessions()
+  private void enqueueSessionIfRequired( @Nonnull final ReplicantSession session )
   {
-    if ( !_queue.isEmpty() && LOG.isLoggable( Level.FINEST ) )
+    if ( null == _workStates.putIfAbsent( session.getId(), WorkState.QUEUED ) )
     {
-      LOG.log( Level.FINEST, () -> "processPendingSessions() invoked with queue.size=" + _queue.size() );
-    }
-    try
-    {
-      ReplicantSession session;
-      while ( null != ( session = _queue.poll( QUEUE_TIMEOUT, TimeUnit.MILLISECONDS ) ) )
-      {
-        processPendingSession( session );
-      }
-    }
-    catch ( final InterruptedException ignored )
-    {
-    }
-    catch ( final Throwable t )
-    {
-      LOG.log( Level.SEVERE, t, () -> "Error in processPendingSessions" );
+      _queue.add( session );
     }
   }
 
-  private void processPendingSession( @Nonnull final ReplicantSession session )
+  private void scheduleDrainTasks()
   {
-    final var id = session.getId();
-    LOG.log( Level.FINEST, () -> "Processing pending ChangeSets for session " + session.getId() );
-    if ( session.isOpen() && !_inProgress.containsKey( id ) )
+    if ( _stopping )
     {
-      final var lock = session.getLock();
+      return;
+    }
+    while ( !_queue.isEmpty() )
+    {
+      if ( !reserveDrainTask() )
+      {
+        return;
+      }
       try
       {
-        lock.lockInterruptibly();
-        _inProgress.put( id, session );
-        Packet packet;
-        while ( null != ( packet = session.popPendingPacket() ) )
-        {
-          _sessionManager.sendChangeMessage( session, packet );
-        }
+        submitDrainTask( this::runDrainTask );
       }
-      catch ( final InterruptedException ignored )
+      catch ( final RuntimeException e )
       {
-        LOG.log( Level.FINEST, () -> "Error completing send of packet " + session.getId() );
-
-        session.closeDueToInterrupt();
-      }
-      finally
-      {
-        //noinspection resource
-        _inProgress.remove( id );
-        lock.unlock();
+        _activeDrainTasks.decrementAndGet();
+        LOG.log( Level.SEVERE,
+                 e,
+                 () -> "Unable to submit Replicant drain task. queue.size=" + _queue.size() +
+                       " activeDrainTasks=" + _activeDrainTasks.get() );
+        scheduleDelayedRetry();
+        return;
       }
     }
+  }
+
+  private boolean reserveDrainTask()
+  {
+    while ( true )
+    {
+      final var activeDrainTasks = _activeDrainTasks.get();
+      if ( activeDrainTasks >= _maxConcurrentDrainTasks || activeDrainTasks >= _workStates.size() )
+      {
+        return false;
+      }
+      if ( _activeDrainTasks.compareAndSet( activeDrainTasks, activeDrainTasks + 1 ) )
+      {
+        return true;
+      }
+    }
+  }
+
+  private void runDrainTask()
+  {
+    var madeProgress = false;
+    try
+    {
+      madeProgress = drainQueuedSessions();
+    }
+    catch ( final Throwable t )
+    {
+      LOG.log( Level.SEVERE, t, () -> "Error in Replicant drain task" );
+    }
+    finally
+    {
+      _activeDrainTasks.decrementAndGet();
+      if ( !_stopping && !_queue.isEmpty() )
+      {
+        if ( madeProgress )
+        {
+          scheduleDrainTasks();
+        }
+        else
+        {
+          scheduleDelayedRetry();
+        }
+      }
+    }
+  }
+
+  private boolean drainQueuedSessions()
+  {
+    var madeProgress = false;
+    final var processedSessionIds = new HashSet<String>();
+    for ( var i = 0; i < _maxSessionsPerDrainTask; i++ )
+    {
+      final var session = _queue.poll();
+      if ( null == session )
+      {
+        break;
+      }
+      if ( !processedSessionIds.add( session.getId() ) )
+      {
+        _queue.add( session );
+        break;
+      }
+      if ( processPendingSession( session ) )
+      {
+        madeProgress = true;
+      }
+    }
+    return madeProgress;
+  }
+
+  private boolean processPendingSession( @Nonnull final ReplicantSession session )
+  {
+    final var id = session.getId();
+    LOG.log( Level.FINEST, () -> "Processing pending ChangeSets for session " + id );
+    if ( !_workStates.replace( id, WorkState.QUEUED, WorkState.RUNNING ) )
+    {
+      return false;
+    }
+    if ( !session.isOpen() )
+    {
+      _workStates.remove( id, WorkState.RUNNING );
+      return true;
+    }
+    final var lock = session.getLock();
+    if ( !lock.tryLock() )
+    {
+      requeueRunningSession( session );
+      return false;
+    }
+    var processedPacket = false;
+    var closeSession = false;
+    try
+    {
+      for ( var i = 0; i < _maxPacketsPerRun; i++ )
+      {
+        final var packet = session.popPendingPacket();
+        if ( null == packet )
+        {
+          break;
+        }
+        processedPacket = true;
+        _sessionManager.sendChangeMessage( session, packet );
+        if ( !session.isOpen() )
+        {
+          closeSession = true;
+          break;
+        }
+      }
+    }
+    catch ( final Throwable t )
+    {
+      LOG.log( Level.SEVERE, t, () -> "Error completing send of packet for session " + id );
+      session.close( new CloseReason( CloseReason.CloseCodes.UNEXPECTED_CONDITION, "Packet processing failed" ) );
+      closeSession = true;
+    }
+    finally
+    {
+      lock.unlock();
+    }
+    if ( closeSession || !session.isOpen() )
+    {
+      _workStates.remove( id, WorkState.RUNNING );
+      return true;
+    }
+    _workStates.remove( id, WorkState.RUNNING );
+    if ( session.hasPendingPackets() )
+    {
+      enqueueSessionIfRequired( session );
+    }
+    return processedPacket;
+  }
+
+  private void requeueRunningSession( @Nonnull final ReplicantSession session )
+  {
+    if ( _workStates.replace( session.getId(), WorkState.RUNNING, WorkState.QUEUED ) )
+    {
+      _queue.add( session );
+    }
+  }
+
+  private void scheduleDelayedRetry()
+  {
+    if ( _stopping || _queue.isEmpty() || !_retryScheduled.compareAndSet( false, true ) )
+    {
+      return;
+    }
+    try
+    {
+      scheduleRetryTask( this::runDelayedRetry );
+    }
+    catch ( final RuntimeException e )
+    {
+      _retryScheduled.set( false );
+      LOG.log( Level.SEVERE,
+               e,
+               () -> "Unable to schedule Replicant drain retry. queue.size=" + _queue.size() +
+                     " activeDrainTasks=" + _activeDrainTasks.get() );
+    }
+  }
+
+  @VisibleForTesting
+  void runDelayedRetry()
+  {
+    _retryScheduled.set( false );
+    if ( !_stopping && !_queue.isEmpty() )
+    {
+      scheduleDrainTasks();
+    }
+  }
+
+  @VisibleForTesting
+  void submitDrainTask( @Nonnull final Runnable task )
+  {
+    _executor.execute( task );
+  }
+
+  @VisibleForTesting
+  void scheduleRetryTask( @Nonnull final Runnable task )
+  {
+    _executor.schedule( task, RETRY_DELAY, TimeUnit.MILLISECONDS );
+  }
+
+  @VisibleForTesting
+  int getActiveDrainTaskCount()
+  {
+    return _activeDrainTasks.get();
+  }
+
+  @VisibleForTesting
+  int getQueuedSessionCount()
+  {
+    return _queue.size();
+  }
+
+  @VisibleForTesting
+  int getWorkStateCount()
+  {
+    return _workStates.size();
+  }
+
+  @VisibleForTesting
+  boolean isRetryScheduled()
+  {
+    return _retryScheduled.get();
+  }
+
+  @VisibleForTesting
+  int getMaxConcurrentDrainTasks()
+  {
+    return _maxConcurrentDrainTasks;
+  }
+
+  @VisibleForTesting
+  int getMaxPacketsPerRun()
+  {
+    return _maxPacketsPerRun;
+  }
+
+  @VisibleForTesting
+  int getMaxSessionsPerDrainTask()
+  {
+    return _maxSessionsPerDrainTask;
+  }
+
+  @VisibleForTesting
+  void setMaxConcurrentDrainTasks( final int maxConcurrentDrainTasks )
+  {
+    _maxConcurrentDrainTasks = maxConcurrentDrainTasks;
+  }
+
+  @SuppressWarnings( "SameParameterValue" )
+  @VisibleForTesting
+  void setMaxPacketsPerRun( final int maxPacketsPerRun )
+  {
+    _maxPacketsPerRun = maxPacketsPerRun;
+  }
+
+  @SuppressWarnings( "SameParameterValue" )
+  @VisibleForTesting
+  void setMaxSessionsPerDrainTask( final int maxSessionsPerDrainTask )
+  {
+    _maxSessionsPerDrainTask = maxSessionsPerDrainTask;
+  }
+
+  private enum WorkState
+  {
+    QUEUED,
+    RUNNING
   }
 }